refactoring hud/sprites/game
