+++ TODO.md
@@ -15,7 +15,7 @@
 * [ ] high score storage support (per-script in localStorage)
 * [ ] add ellipse drawing (filled and empty)
 * [x] add empty rectangle drawing (using lines)
-* [ ] think about frame handler order (see `sprite.ts` for a list)
+* [ ] think about frame handler order (see `sprites.ts` for a list)
 * [ ] think if we want to support devices with very low memory (eg SAMD21 has 32k and one screen is 8k; we need double buffering, and also some other random memory which leaves around 1 screen worth of memory for the user)
 * [ ] nice design for GAME OVER - probably should avoid text - localization
 * [ ] nice design for splash
